+++ NEWS.md
@@ -1,6 +1,7 @@
 # tidyterra (development version)
 
 -   Now `geom_spatraster_rgb()` works with `facet_wrap()` (#35)
+-   Improve facetting when the plot facets are created using non-Spat\* layers.
 
 # tidyterra 0.2.0